Testing Version of Site
Test out the different files. Each link enters the information into the database with the given constraints. An error message will pop up if there is a constraint issue... If the input doesn't work for you, uncomment the if else statemen below that I have in most of the files.    It will tell you exactly what went wrong with input , but everything should be working fine.
                
             /*if (!mysqli_query($conn,$sql)) {
                die('Error: ' . mysqli_error($conn));     // Checks if insert into the database was succesful
                } */